+++ pygeos/tests/test_constructive.py
@@ -3,6 +3,7 @@ import pytest
 
 import pygeos
 from pygeos import Geometry, GEOSException
+from pygeos.testing import assert_geometries_equal
 
 from .common import (
     all_types,
@@ -271,9 +272,12 @@ def test_offset_curve_join_style_invalid():
                 "POLYGON((0 0, 0 10, 10 10, 10 0, 0 0), (2 2, 4 2, 4 4, 2 4, 2 2))"
             ),
         ),
-        (
+        pytest.param(
             pygeos.Geometry("MULTILINESTRING ((0 0, 1 2), (3 3, 4 4))"),
             pygeos.Geometry("MULTILINESTRING ((1 2, 0 0), (4 4, 3 3))"),
+            marks=pytest.mark.skipif(
+                pygeos.geos_version < (3, 8, 0), reason="GEOS < 3.8"
+            ),
         ),
         (
             pygeos.Geometry(
@@ -295,7 +299,7 @@ def test_offset_curve_join_style_invalid():
     ],
 )
 def test_reverse(geom, expected):
-    assert pygeos.equals(pygeos.reverse(geom), expected)
+    assert_geometries_equal(pygeos.reverse(geom), expected)
 
 
 @pytest.mark.skipif(pygeos.geos_version < (3, 7, 0), reason="GEOS < 3.7")
@@ -307,7 +311,7 @@ def test_reverse_none():
     expected = pygeos.Geometry("POLYGON ((0 0,  0 1, 1 1, 1 0, 0 0))")
     result = pygeos.reverse([None, geometry])
     assert result[0] is None
-    assert pygeos.equals(result[1], expected)
+    assert_geometries_equal(result[1], expected)
 
 
 @pytest.mark.skipif(pygeos.geos_version < (3, 7, 0), reason="GEOS < 3.7")
@@ -339,7 +343,7 @@ def test_reverse_invalid_type(geom):
 def test_clip_by_rect(geom, expected):
     geom, expected = pygeos.Geometry(geom), pygeos.Geometry(expected)
     actual = pygeos.clip_by_rect(geom, 10, 10, 20, 20)
-    assert pygeos.equals(actual, expected)
+    assert_geometries_equal(actual, expected)
 
 
 @pytest.mark.parametrize(
@@ -374,7 +378,7 @@ def test_clip_by_rect(geom, expected):
 def test_clip_by_rect_polygon(geom, rect, expected):
     geom, expected = pygeos.Geometry(geom), pygeos.Geometry(expected)
     actual = pygeos.clip_by_rect(geom, *rect)
-    assert pygeos.equals(actual, expected)
+    assert_geometries_equal(actual, expected)
 
 
 @pytest.mark.parametrize("geometry", all_types)
@@ -605,14 +609,14 @@ def test_segmentize_tolerance_nan(geometry):
 )
 def test_segmentize_empty(geometry):
     actual = pygeos.segmentize(geometry, tolerance=5)
-    assert pygeos.equals(actual, geometry).all()
+    assert_geometries_equal(actual, geometry)
 
 
 @pytest.mark.skipif(pygeos.geos_version < (3, 10, 0), reason="GEOS < 3.10")
 @pytest.mark.parametrize("geometry", [point, point_z, multi_point])
 def test_segmentize_no_change(geometry):
     actual = pygeos.segmentize(geometry, tolerance=5)
-    assert pygeos.equals(actual, geometry).all()
+    assert_geometries_equal(actual, geometry)
 
 
 @pytest.mark.skipif(pygeos.geos_version < (3, 10, 0), reason="GEOS < 3.10")
@@ -686,7 +690,7 @@ def test_segmentize_none():
 )
 def test_segmentize(geometry, tolerance, expected):
     actual = pygeos.segmentize(geometry, tolerance)
-    assert pygeos.equals(actual, geometry).all()
+    assert_geometries_equal(actual, expected)
 
 
 @pytest.mark.skipif(pygeos.geos_version < (3, 8, 0), reason="GEOS < 3.8")
@@ -728,7 +732,7 @@ def test_minimum_bounding_circle_all_types(geometry):
 )
 def test_minimum_bounding_circle(geometry, expected):
     actual = pygeos.minimum_bounding_circle(geometry)
-    assert pygeos.equals(actual, expected).all()
+    assert_geometries_equal(actual, expected)
 
 
 @pytest.mark.skipif(pygeos.geos_version < (3, 6, 0), reason="GEOS < 3.6")