+++ pygeos/tests/test_linear.py
@@ -2,6 +2,7 @@ import numpy as np
 import pytest
 
 import pygeos
+from pygeos.testing import assert_geometries_equal
 
 from .common import (
     empty_line_string,
@@ -20,9 +21,9 @@ def test_line_interpolate_point_geom_array():
     actual = pygeos.line_interpolate_point(
         [line_string, linear_ring, multi_line_string], -1
     )
-    assert pygeos.equals(actual[0], pygeos.Geometry("POINT (1 0)"))
-    assert pygeos.equals(actual[1], pygeos.Geometry("POINT (0 1)"))
-    assert pygeos.equals_exact(
+    assert_geometries_equal(actual[0], pygeos.Geometry("POINT (1 0)"))
+    assert_geometries_equal(actual[1], pygeos.Geometry("POINT (0 1)"))
+    assert_geometries_equal(
         actual[2], pygeos.Geometry("POINT (0.5528 1.1056)"), tolerance=0.001
     )
 
@@ -31,16 +32,16 @@ def test_line_interpolate_point_geom_array_normalized():
     actual = pygeos.line_interpolate_point(
         [line_string, linear_ring, multi_line_string], 1, normalized=True
     )
-    assert pygeos.equals(actual[0], pygeos.Geometry("POINT (1 1)"))
-    assert pygeos.equals(actual[1], pygeos.Geometry("POINT (0 0)"))
-    assert pygeos.equals(actual[2], pygeos.Geometry("POINT (1 2)"))
+    assert_geometries_equal(actual[0], pygeos.Geometry("POINT (1 1)"))
+    assert_geometries_equal(actual[1], pygeos.Geometry("POINT (0 0)"))
+    assert_geometries_equal(actual[2], pygeos.Geometry("POINT (1 2)"))
 
 
 def test_line_interpolate_point_float_array():
     actual = pygeos.line_interpolate_point(line_string, [0.2, 1.5, -0.2])
-    assert pygeos.equals(actual[0], pygeos.Geometry("POINT (0.2 0)"))
-    assert pygeos.equals(actual[1], pygeos.Geometry("POINT (1 0.5)"))
-    assert pygeos.equals(actual[2], pygeos.Geometry("POINT (1 0.8)"))
+    assert_geometries_equal(actual[0], pygeos.Geometry("POINT (0.2 0)"))
+    assert_geometries_equal(actual[1], pygeos.Geometry("POINT (1 0.5)"))
+    assert_geometries_equal(actual[2], pygeos.Geometry("POINT (1 0.8)"))
 
 
 @pytest.mark.parametrize("normalized", [False, True])
@@ -59,7 +60,7 @@ def test_line_interpolate_point_empty(geom, normalized):
     # These geometries segfault in some versions of GEOS (in 3.8.0, still
     # some of them segfault). Instead, we patched this to return POINT EMPTY.
     # This matches GEOS 3.8.0 behavior on simple empty geometries.
-    assert pygeos.equals(
+    assert_geometries_equal(
         pygeos.line_interpolate_point(geom, 0.2, normalized=normalized), empty_point
     )
 
@@ -132,15 +133,17 @@ def test_line_locate_point_invalid_geometry(normalized):
 
 def test_line_merge_geom_array():
     actual = pygeos.line_merge([line_string, multi_line_string])
-    assert pygeos.equals(actual[0], line_string)
-    assert pygeos.equals(actual[1], multi_line_string)
+    assert_geometries_equal(actual[0], line_string)
+    assert_geometries_equal(actual[1], pygeos.Geometry("LINESTRING (0 0, 1 2)"))
 
 
 def test_shared_paths_linestring():
     g1 = pygeos.linestrings([(0, 0), (1, 0), (1, 1)])
     g2 = pygeos.linestrings([(0, 0), (1, 0)])
     actual1 = pygeos.shared_paths(g1, g2)
-    assert pygeos.equals(pygeos.get_geometry(actual1, 0), g2)
+    assert_geometries_equal(
+        pygeos.get_geometry(actual1, 0), pygeos.multilinestrings([g2])
+    )
 
 
 def test_shared_paths_none():