+++ pygeos/tests/test_set_operations.py
@@ -4,6 +4,7 @@ import pytest
 import pygeos
 from pygeos import Geometry
 from pygeos.decorators import UnsupportedGEOSOperation
+from pygeos.testing import assert_geometries_equal
 
 from .common import all_types, multi_polygon, point, polygon
 
@@ -122,7 +123,7 @@ def test_set_operation_reduce_one_none(func, related_func, none_position):
     test_data.insert(none_position, None)
     actual = func(test_data)
     expected = related_func(reduce_test_data[0], reduce_test_data[1])
-    assert pygeos.equals(actual, expected)
+    assert_geometries_equal(actual, expected)
 
 
 @pytest.mark.parametrize("none_position", range(3))
@@ -133,7 +134,7 @@ def test_set_operation_reduce_two_none(func, related_func, none_position):
     test_data.insert(none_position, None)
     actual = func(test_data)
     expected = related_func(reduce_test_data[0], reduce_test_data[1])
-    assert pygeos.equals(actual, expected)
+    assert_geometries_equal(actual, expected)
 
 
 @pytest.mark.parametrize("n", range(1, 3))
@@ -214,7 +215,7 @@ def test_set_operation_prec_reduce_one_none(func, related_func, none_position):
     test_data.insert(none_position, None)
     actual = func(test_data, grid_size=1)
     expected = related_func(reduce_test_data[0], reduce_test_data[1], grid_size=1)
-    assert pygeos.equals(actual, expected)
+    assert_geometries_equal(actual, expected)
 
 
 @pytest.mark.skipif(pygeos.geos_version < (3, 9, 0), reason="GEOS < 3.9")
@@ -226,7 +227,7 @@ def test_set_operation_prec_reduce_two_none(func, related_func, none_position):
     test_data.insert(none_position, None)
     actual = func(test_data, grid_size=1)
     expected = related_func(reduce_test_data[0], reduce_test_data[1], grid_size=1)
-    assert pygeos.equals(actual, expected)
+    assert_geometries_equal(actual, expected)
 
 
 @pytest.mark.skipif(pygeos.geos_version < (3, 9, 0), reason="GEOS < 3.9")
@@ -252,7 +253,7 @@ def test_coverage_union_reduce_1dim(n):
     expected = test_data[0]
     for i in range(1, n):
         expected = pygeos.coverage_union(expected, test_data[i])
-    assert pygeos.equals(actual, expected)
+    assert_geometries_equal(actual, expected, normalize=True)
 
 
 @pytest.mark.skipif(pygeos.geos_version < (3, 8, 0), reason="GEOS < 3.8")